+++ adapted/config/base.py
@@ -9,10 +9,12 @@ Contact: w.vandertoorn@fu-berlin.de
 import pprint
 from dataclasses import asdict, dataclass, fields
 from pathlib import Path
-from typing import Tuple, Type, TypeVar, Union, get_args
+from typing import Any, MutableMapping, Tuple, Type, TypeVar, Union, get_args
+import logging
 
 import numpy as np
 import toml
+import sys
 
 
 def get_field_type_annotation(field):
@@ -73,12 +75,12 @@ class BaseConfig:
 
 @dataclass
 class NestedConfig(BaseConfig):
-    def pretty_print(self):
+    def pretty_print(self, file=sys.stdout):
         for key, val in self.dict().items():
             if isinstance(val, BaseConfig):
-                print(f"{key}:\n{val.pretty_print()}")
+                print(f"{key}:\n{val.pretty_print()}", file=file)
             else:
-                print(f"{key}: {val}")
+                print(f"{key}: {val}", file=file)
 
     def typed_dict(self):
         return {
@@ -94,29 +96,38 @@ def load_nested_config_from_file(
     file_path: Union[str, Path], config_class: Type[T] = NestedConfig
 ) -> T:
     config_file = toml.load(file_path)
-    config_file_keys = set(config_file.keys())
+    return nested_config_from_dict(config_file, config_class)
+
+
+def nested_config_from_dict(
+    config_dict: Union[dict, MutableMapping[str, Any]],
+    config_class: Type[T] = NestedConfig,
+) -> T:
+    config_file_keys = set(config_dict.keys())
 
     valid_keys = set([f.name for f in fields(config_class)])
     unknown_keys = [key for key in config_file_keys if key not in valid_keys]
 
     if unknown_keys:
-        raise ValueError(
-            f"Invalid config file. Unknown key(s): {', '.join(unknown_keys)}. "
-            f"Valid keys are: {', '.join(valid_keys)}"
-        )
+        msg = f"Invalid config file. Unknown key(s): {', '.join(unknown_keys)}. "
+        msg += f"Valid keys are: {', '.join(valid_keys)}"
+        logging.error(msg)
+        raise ValueError(msg)
 
     config_obj = config_class()
 
     # First, handle entries without a section header
     for key in config_file_keys:
-        if not isinstance(config_file[key], dict):
+        if not isinstance(config_dict[key], dict):
             if key in valid_keys:
-                setattr(config_obj, key, config_file[key])
+                setattr(config_obj, key, config_dict[key])
             else:
-                raise ValueError(f"Invalid config file. Unknown key: {key}")
+                msg = f"Invalid config file. Unknown key: {key}"
+                logging.error(msg)
+                raise ValueError(msg)
 
     # Then, handle sections
-    for section, content in config_file.items():
+    for section, content in config_dict.items():
         if isinstance(content, dict):
             if section in valid_keys:
                 field_type = next(
@@ -124,21 +135,22 @@ def load_nested_config_from_file(
                     for f in fields(config_class)
                     if f.name == section
                 )
-                if issubclass(field_type, BaseConfig):
-                    pass
+                if issubclass(field_type, BaseConfig):  # type: ignore
                     try:
                         setattr(config_obj, section, field_type(**content))
                     except:
-                        raise ValueError(
-                            f"Invalid config file. Could not parse section {section}"
-                            f" with content {content}"
-                            f" as {field_type}"
-                        )
+                        msg = f"Invalid config file. Could not parse section {section}"
+                        msg += f" with content {content}"
+                        msg += f" as {field_type}"
+                        logging.error(msg)
+                        raise ValueError(msg)
                 else:
-                    raise ValueError(
-                        f"Invalid section type for {section}: {field_type}"
-                    )
+                    msg = f"Invalid section type for {section}: {field_type}"
+                    logging.error(msg)
+                    raise ValueError(msg)
             else:
-                raise ValueError(f"Invalid config file. Unknown section: {section}")
+                msg = f"Invalid config file. Unknown section: {section}"
+                logging.error(msg)
+                raise ValueError(msg)
 
     return config_obj