+++ adapted/config/file_proc.py
@@ -19,6 +19,8 @@ class BatchConfig(BaseConfig):
     num_proc: int = -1  # default to number of cores
     batch_size: int = 4000
     minibatch_size: int = 50
+    bidx_passed: int = 0
+    bidx_failed: int = 0
 
 
 @dataclass
@@ -32,6 +34,7 @@ class OutputConfig(BaseConfig):
 class InputConfig(BaseConfig):
     files: List[str] = field(default_factory=list)
     read_ids: Union[List[str], np.ndarray] = field(default_factory=list)
+    continue_from: str = field(default="")
 
 
 @dataclass